played with parameters, made Incubator Observable
In preparation for method of use:

There will be a single test java file that will run the experiment.

The important class here is the ClassifierIncubator

the incubator works by calling the generateReasoningEngine() function

this can be called in two ways
1 - generateReasoningEngine(File learningData, float fitnessTreshold)
2 - generateReasoningEngine(int maxLeafCount, int populationSize, File
learningData, float fitnessTreshold)

print statements have yet to be cleaned (much as the entire code tbh coz
I had to hurry to hand it in). Will be replaced with "notify" for the
observers